refactor(user): share authenticated payload mappers
